+++ app/routes/components/Modal/changelog.codegen.ts
@@ -6,4 +6,6 @@ export const changelog = ''
   + '- 2025-10-14: Make `Modal.Content` support `center` and `shadow` props. Update\n'
   + '  `Modal.Content` styles.\n'
   + '- 2025-10-27: Add `Modal.Raw` component.\n'
+  + '- 2025-11-23: Remove the dependency on `ThemeProvider`. Support `yOffset` prop\n'
+  + '  for `Modal.Content`.\n'
   ;