+++ src/components/Modal/index.ts
@@ -1,3 +1,3 @@
-// Modal from copy-ui @ 2025-10-27.
+// Modal from copy-ui @ 2025-11-23.
 
 export { Modal } from "./modal";