+++ src/components/Modal/CHANGELOG.md
@@ -2,3 +2,5 @@
 - 2025-10-14: Make `Modal.Content` support `center` and `shadow` props. Update
   `Modal.Content` styles.
 - 2025-10-27: Add `Modal.Raw` component.
+- 2025-11-23: Remove the dependency on `ThemeProvider`. Support `yOffset` prop
+  for `Modal.Content`.